+++ android/src/org/coolreader/crengine/AboutDialog.java
@@ -21,12 +21,8 @@
 
 package org.coolreader.crengine;
 
-import android.content.ActivityNotFoundException;
-import android.content.Intent;
-import android.net.Uri;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.widget.Button;
 import android.widget.TabHost;
 import android.widget.TabHost.TabContentFactory;
 import android.widget.TextView;
@@ -36,7 +32,6 @@ import org.coolreader.R;
 
 import java.util.ArrayList;
 import java.util.Iterator;
-import java.util.Random;
 
 public class AboutDialog extends BaseDialog implements TabContentFactory {
 	final CoolReader mCoolReader;
@@ -44,52 +39,6 @@ public class AboutDialog extends BaseDialog implements TabContentFactory {
 	private View mAppTab;
 	private View mDirsTab;
 	private View mLicenseTab;
-	private View mDonationTab;
-
-	private boolean isPackageInstalled( String packageName ) {
-		try {
-			mCoolReader.getPackageManager().getApplicationInfo(packageName, 0);
-			return true;
-		} catch ( Exception e ) {
-			return false;
-		}
-	}
-
-	private void installPackage( String packageName ) {
-		Log.i("cr3", "installPackageL " + packageName);
-		try {
-			mCoolReader.startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse("market://details?id=" + packageName)));
-		} catch ( ActivityNotFoundException e ) {
-			mCoolReader.showToast("Cannot run Android Market application");
-		}
-	}
-	
-	private void setupDonationButton( final Button btn, final String packageName ) {
-		if ( isPackageInstalled(packageName)) {
-			btn.setEnabled(false);
-			btn.setText(R.string.dlg_about_donation_installed);
-		} else {
-			btn.setOnClickListener(v -> installPackage(packageName));
-		}
-	}
-	
-	private void setupInAppDonationButton( final Button btn, final double amount ) {
-		btn.setText("$" + amount);
-		btn.setOnClickListener(v -> mCoolReader.makeDonation(amount));
-	}
-	
-	private void updateTotalDonations() {
-		double amount = mCoolReader.getTotalDonations();
-		if (isPackageInstalled("org.coolreader.donation.gold"))
-			amount += 10.0;
-		if (isPackageInstalled("org.coolreader.donation.silver"))
-			amount += 3.0;
-		if (isPackageInstalled("org.coolreader.donation.bronze"))
-			amount += 1.0;
-		TextView text = ((TextView)mDonationTab.findViewById(R.id.btn_about_donation_total));
-		if (text != null)
-			text.setText(mCoolReader.getString(R.string.dlg_about_donation_total) + " $" + amount);
-	}
 
 	public AboutDialog( CoolReader activity)
 	{
@@ -159,24 +108,6 @@ public class AboutDialog extends BaseDialog implements TabContentFactory {
 		mLicenseTab = inflater.inflate(R.layout.about_dialog_license, null);
 		String license = Engine.getInstance(mCoolReader).loadResourceUtf8(R.raw.license);
 		((TextView)mLicenseTab.findViewById(R.id.license)).setText(license);
-		boolean billingSupported = mCoolReader.isDonationSupported();
-		mDonationTab = inflater.inflate(billingSupported ? R.layout.about_dialog_donation2 : R.layout.about_dialog_donation, null);
-
-		if (billingSupported) {
-			setupInAppDonationButton(mDonationTab.findViewById(R.id.btn_about_donation_install_vip), 100);
-			setupInAppDonationButton(mDonationTab.findViewById(R.id.btn_about_donation_install_platinum), 30);
-			setupInAppDonationButton(mDonationTab.findViewById(R.id.btn_about_donation_install_gold), 10);
-			setupInAppDonationButton(mDonationTab.findViewById(R.id.btn_about_donation_install_silver), 3);
-			setupInAppDonationButton(mDonationTab.findViewById(R.id.btn_about_donation_install_bronze), 1);
-			setupInAppDonationButton(mDonationTab.findViewById(R.id.btn_about_donation_install_iron), 0.3);
-			updateTotalDonations();
-			mCoolReader.setDonationListener(total -> updateTotalDonations());
-			setOnDismissListener(dialog -> mCoolReader.setDonationListener(null));
-		} else {
-			setupDonationButton(mDonationTab.findViewById(R.id.btn_about_donation_install_gold), "org.coolreader.donation.gold");
-			setupDonationButton(mDonationTab.findViewById(R.id.btn_about_donation_install_silver), "org.coolreader.donation.silver");
-			setupDonationButton(mDonationTab.findViewById(R.id.btn_about_donation_install_bronze), "org.coolreader.donation.bronze");
-		}
 		
 		tabs.setup();
 		TabHost.TabSpec tsApp = tabs.newTabSpec("App");
@@ -197,20 +128,9 @@ public class AboutDialog extends BaseDialog implements TabContentFactory {
 		tsLicense.setContent(this);
 		tabs.addTab(tsLicense);
 		
-		TabHost.TabSpec tsDonation = tabs.newTabSpec("Donation");
-		tsDonation.setIndicator("", 
-				getContext().getResources().getDrawable(R.drawable.ic_menu_emoticons));
-		tsDonation.setContent(this);
-		tabs.addTab(tsDonation);
-		
 		setView( tabs );
-
-		// 25% chance to show Donations tab
-		if ((rnd.nextInt() & 3) == 3)
-			tabs.setCurrentTab(3);
 		
 	}
-	private static Random rnd = new Random(android.os.SystemClock.uptimeMillis()); 
 
 	
 	@Override
@@ -222,8 +142,6 @@ public class AboutDialog extends BaseDialog implements TabContentFactory {
 			return mDirsTab;
 		else if ( "License".equals(tag) )
 			return mLicenseTab;
-		else if ( "Donation".equals(tag) )
-			return mDonationTab;
 		return null;
 	}